Add stable device index via FNV-1a hash; fix SNMP iterator for large indices
Derive smartmonDeviceIndex from FNV-1a32(serial + '|' + model) so the index
is stable across daemon restarts instead of being assigned sequentially.

Two fixes required to make large hash indices visible in SNMP walks:

- Cap hash to 31 bits (& 0x7FFFFFFF): net-snmp's table iterator silently
  drops GETNEXT results for OID sub-identifiers >= 2^31.  31 bits gives
  ~2 billion unique indices, ample for any realistic drive count.

- Use uint32_t (4 bytes) instead of u_long (8 bytes) when calling
  snmp_set_var_typed_value with ASN_UNSIGNED in all get_next functions.
  Passing sizeof(u_long)=8 for a 32-bit SNMP type caused incorrect index
  storage in net-snmp's internal variable list.

- Sort g_cache.devices by index after initial scan to match net-snmp's
  expected iteration order.

upsert_device() gains a hint_idx parameter; callers pass the FNV hash as a
starting point with collision fallback to hint_idx+1, hint_idx+2, etc.

+++ src/snmp_nvme_mib.cpp
@@ -80,7 +80,7 @@ nvme_health_get_next(void **loop_ctx, void **data_ctx,
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
     // Index 1: smartmonDeviceIndex
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
     // Index 2: smartmonNvmeHealthIndex (always 1 per device)
     v = 1;
@@ -167,9 +167,9 @@ nvme_st_get_next(void **loop_ctx, void **data_ctx,
     CacheNvmeSelfTestRow &row = g_cache.nvme_selftests[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
-    v = (u_long)row.entry_index;
+    v = row.entry_index;
     snmp_set_var_typed_value(put_idx->next_variable, ASN_UNSIGNED,
                              (u_char*)&v, sizeof(v));
     return put_idx;
@@ -250,7 +250,7 @@ nvme_ctrl_get_next(void **loop_ctx, void **data_ctx,
     CacheNvmeControllerRow &row = g_cache.nvme_controllers[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
     v = 1;  // controllerIndex = 1 per device
     snmp_set_var_typed_value(put_idx->next_variable, ASN_UNSIGNED,
@@ -343,9 +343,9 @@ nvme_ns_get_next(void **loop_ctx, void **data_ctx,
     CacheNvmeNamespaceRow &row = g_cache.nvme_namespaces[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
-    v = (u_long)row.namespace_id;
+    v = row.namespace_id;
     snmp_set_var_typed_value(put_idx->next_variable, ASN_UNSIGNED,
                              (u_char*)&v, sizeof(v));
     return put_idx;
@@ -415,9 +415,9 @@ nvme_el_get_next(void **loop_ctx, void **data_ctx,
     CacheNvmeErrorLogRow &row = g_cache.nvme_error_log[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
-    v = (u_long)row.entry_index;
+    v = row.entry_index;
     snmp_set_var_typed_value(put_idx->next_variable, ASN_UNSIGNED,
                              (u_char*)&v, sizeof(v));
     return put_idx;
@@ -505,7 +505,7 @@ nvme_cap_get_next(void **loop_ctx, void **data_ctx,
     CacheNvmeCapabilityRow &row = g_cache.nvme_capabilities[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
     v = 1;  // capabilityIndex = 1 per device
     snmp_set_var_typed_value(put_idx->next_variable, ASN_UNSIGNED,
@@ -585,9 +585,9 @@ nvme_ps_get_next(void **loop_ctx, void **data_ctx,
     CacheNvmePowerStateRow &row = g_cache.nvme_power_states[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
-    v = (u_long)row.state_index;
+    v = row.state_index;
     snmp_set_var_typed_value(put_idx->next_variable, ASN_UNSIGNED,
                              (u_char*)&v, sizeof(v));
     return put_idx;
@@ -670,12 +670,12 @@ nvme_lba_get_next(void **loop_ctx, void **data_ctx,
     CacheNvmeLbaFormatRow &row = g_cache.nvme_lba_formats[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long v = (u_long)row.device_index;
+    uint32_t v = row.device_index;
     snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&v, sizeof(v));
-    v = (u_long)row.namespace_id;
+    v = row.namespace_id;
     snmp_set_var_typed_value(put_idx->next_variable, ASN_UNSIGNED,
                              (u_char*)&v, sizeof(v));
-    v = (u_long)row.format_id;
+    v = row.format_id;
     snmp_set_var_typed_value(put_idx->next_variable->next_variable, ASN_UNSIGNED,
                              (u_char*)&v, sizeof(v));
     return put_idx;
